[fix] 기능에 맞게 리팩터링
what :
- admin, gamer, course 서비스 기능 각 기능에 맞게 리팩터링

why : 
- 서비스의 목적에 맞게 기능 분배
