+++ Backend/src/main/java/com/github/backend/web/CourseController.java
@@ -17,56 +17,12 @@ public class CourseController {
 
     private final CourseService courseService;
 
-    @PostMapping("/register")
-    public RegisterCourse.Response registerCourse(
-            @RequestBody @Valid RegisterCourse.Request request
-    ){
-        Course course = courseService.registerCourse(
-                request.getGamerId(),
-                request.getTitle(),
-                request.getVideoUrl(),
-                request.getThumbnailUrl(),
-                request.getComment(),
-                request.getLevel(),
-                request.getRace(),
-                request.getPrice()
-        );
-
-        return new RegisterCourse.Response(
-                course.getGamer().getName(),
-                course.getTitle(),
-                course.getRace(),
-                course.getLevel(),
-                course.getComment(),
-                course.getPrice()
-        );
-    }
-
     @GetMapping("/{courseId}")
     public CourseInfoResponse getCourseInfo(
             @PathVariable("courseId") @Valid Long courseId
     ){
         return CourseInfoResponse.from(
                 courseService.getCourseInfo(courseId)
-        );
-    }
-
-    @PutMapping("/{courseId}/edit")
-    public UpdateCourse.Response updateCourseInfo(
-            @PathVariable("courseId") @Valid Long courseId,
-            @RequestBody UpdateCourse.Request request
-    ){
-
-        Course course = courseService.updateCourseInfo(courseId, request);
-        return new UpdateCourse.Response(
-                course.getGamer().getName(),
-                course.getTitle(),
-                course.getVideoUrl(),
-                course.getThumbnailUrl(),
-                course.getComment(),
-                course.getLevel(),
-                course.getRace(),
-                course.getPrice()
         );
     }
